backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.49
pydantic==1.10.11
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
apscheduler==3.9.1
twilio==8.4.1
python-dotenv==1.0.0
jinja2==3.1.2